refactor: extract place-search and routing provider modules
